Check the OS using /etc/<package.conf> instead of lsb_release

+++ setup.sh
@@ -4,7 +4,7 @@ jarvispath=$PWD
 touch jarvis
 chmod +rwx jarvis
 exec 3<> jarvis
-    
+
     echo "#!/bin/bash" >&3
     echo "source $jarvispath/env/bin/activate" >&3
     echo "python $jarvispath/jarviscli/" >&3
@@ -23,12 +23,6 @@ if [[ "$UNAME" == "Darwin" ]]; then
   exit 0
 fi
 
-OS=$(lsb_release -si)
-if [[ "$OS" == "Fedora" ]]; then
-  sudo dnf install ffmpeg
-  sudo dnf install notify-python -y
-  sudo -H pip install -r requirements.txt
-  sudo dnf install python-dbus -y
 # Check if sudo is required
 pip install virtualenv
 if [[ "$?" -eq 2 ]]; then
@@ -38,31 +32,30 @@ fi
 virtualenv env --python=python2.7
 source env/bin/activate
 pip install -r requirements.txt
+
+# Fedora based (>=22)
+if [[ -f "/etc/dnf/dnf.conf" ]]; then
+  sudo dnf install -y ffmpeg python-dbus notify-python
   # chromedriver-install
   wget https://chromedriver.storage.googleapis.com/2.32/chromedriver_linux64.zip
   unzip chromedriver_linux64_2.3.zip
   sudo cp chromedriver /usr/bin/chromedriver
   sudo chown root /usr/bin/chromedriver
   sudo chmod 755 /usr/bin/chromedriver
-elif [[ "$OS" == "Ubuntu" ]] || [[ "$OS" == "LinuxMint" ]]; then
+
+# Debian based
+elif [[ -f "/etc/apt/sources.list" ]]; then
   sudo apt-get install ffmpeg python-imdbpy python-notify2
   sudo apt-get install python-dbus python-dbus-dev libssl-dev libffi-dev libdbus-1-dev libdbus-glib-1-dev
   sudo apt-get install chromium-chromedriver python2.7
-  sudo pip install virtualenv
-  virtualenv env --python=python2.7
-  source env/bin/activate
-  pip install -r requirements.txt
-elif [[ "$OS" == "Kali" ]]; then
-  apt-get install ffmpeg
-  apt-get install python-notify2
-  apt-get install python-dbus
-  pip install -r requirements.txt
-elif [[ "$OS" == "Arch" ]] || [[ "$OS" == "ManjaroLinux" ]]; then
-  sudo pacman -S --noconfirm ffmpeg
-  sudo pacman -S --noconfirm openssl libffi
-  pip install -r requirements.txt
+
+# Arch based
+elif [[ -f "/etc/pacman.conf" ]]; then
+  sudo pacman -S --noconfirm ffmpeg openssl libffi
+
 else
-  echo "Operating System not supported"
+  echo "Operating system not supported"
   exit 1
 fi
+
 exit 0
